Fix build matching and refactor components
