+++ src/PageCharacter/CharacterDisplay/Tabs/TabOptimize/Components/ChartCard/EnhancedPoint.ts
@@ -1,0 +1,11 @@
+export type EnhancedPoint = {
+  x: number
+  y?: number
+  artifactIds: string[]
+  min?: number
+  current?: number
+  highlighted?: number
+}
+export function getEnhancedPointY(pt: EnhancedPoint) {
+  return (pt.y || pt.current || pt.highlighted) as number
+}
